requirements.txt
Django>=5.2.7
djangorestframework>=3.16.1
django-modeltranslation>=0.19.17
django-widget-tweaks>=1.5.0
django-quill-editor>=0.1.42
djangorestframework-simplejwt>=5.5.1
Pillow>=11.3.0
requests>=2.32.5
